fix(checker): refresh rendered diagnostic gates
