+++ crates/tsz-checker/src/error_reporter/mod.rs
@@ -19,6 +19,7 @@ mod assignability_exact_optional;
 mod assignability_helpers;
 mod assignability_keyof_alias_display;
 mod assignability_literal_display;
+mod assignability_missing_property_satisfaction;
 mod assignability_normalized_union;
 mod assignability_numeric_display;
 mod assignability_satisfies;
